Create boxes through a tunnel
You are transporting some boxes through a tunnel, where each box is a parallelepiped, and is characterized by its length, width and height.

The height of the tunnel 41 feet and the width can be assumed to be infinite. A box can be carried through the tunnel only if its height is strictly less than the tunnel's height. Find the volume of each box that can be successfully transported to the other end of the tunnel. Note: Boxes cannot be rotated.

Input Format
The first line contains a single integer n , denoting the number of boxes.
 n lines follow with three integers on each separated by single spaces -  length , width and height,  which are length, width and height in feet of the - i th box.