Create Calculate the nth term
Task
There is a series,sS , where the next term is the sum of previous three terms. Given the first three terms of the series,a ,b ,c and  respectively, you have to output the nth term of the series using recursion.

Input Format
The first line contains a single integer,n .
The next line contains 3 space-separated integers,a ,b , and c.